+++ lib/src/all_web.dart
@@ -1,5 +1,6 @@
 export 'web/delegate.dart';
 export 'web/interpreter.dart';
+export 'web/litertjs_interpreter.dart';
 export 'web/interpreter_options.dart';
 export 'web/isolate_interpreter.dart';
 export 'web/signature_runner.dart';